Tilføjet de andre knapper til ToggleViewList, samt tilføjet kommentarer
