+++ app/src/main/res/values/styles.xml
@@ -26,8 +26,8 @@
 
 
     <style name="DefaultButton">
-        <item name="android:layout_width">60sp</item>
-        <item name="android:layout_height">60sp</item>
+        <item name="android:layout_width">65sp</item>
+        <item name="android:layout_height">65sp</item>
         <item name="android:backgroundTint">@color/grey</item>
         <item name="android:textColor">#2D2F31</item>
         <item name="android:textSize">24sp</item>
